requirements.txt
Flask
python-barcode
Pillow
